Split DrakeCollision.h into separate class files
