+++ systems/plants/constructModelmex.cpp
@@ -275,8 +275,8 @@ void mexFunction( int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[] )
 
       // Set collision filtering bitmasks
       pm = mxGetProperty(pBodies,i,"collision_filter");
-      const uint16_t* group = (uint16_t*)mxGetPr(mxGetField(pm,0,"belongs_to"));
-      const uint16_t* mask = (uint16_t*)mxGetPr(mxGetField(pm,0,"ignores"));
+      const auto group = (DrakeCollision::bitmask*)mxGetPr(mxGetField(pm,0,"belongs_to"));
+      const auto mask = (DrakeCollision::bitmask*)mxGetPr(mxGetField(pm,0,"ignores"));
       //DEBUG
       //cout << "constructModelmex: Group: " << *group << endl;
       //cout << "constructModelmex: Mask " << *mask << endl;
